Separate test classes and tests for better readability
