+++ .gitignore
@@ -7,3 +7,4 @@ node_modules
 .env
 .env.*
 !.env.example
+/src/test-results
