+++ tests/TestHelper/TroiApiStub.js
@@ -1,0 +1,111 @@
+import md5 from "crypto-js/md5.js";
+
+export const correctUser = "user.name";
+export const correctPassword = "s3cr3t";
+
+const mockData = {
+    client: {
+        Name: "DigitalService GmbH des Bundes",
+        Id: 123,
+    },
+    employee: {
+        Id: 456,
+    },
+    calculationPosition: {
+        DisplayPath: "My Project",
+        Id: 789,
+    },
+};
+
+export default class TroiApiStub {
+
+    constructor() {
+        this.entries = [];
+
+        this.correctAuthnHeader = `Basic ${btoa(
+            `${correctUser}:${md5(correctPassword)}`
+        )}`;
+    }
+
+    addEntry(entry) {
+        this.entries.push(entry);
+    }
+
+    deleteEntry(id) {
+        this.entries = this.entries.filter((entry) => entry.id !== id);
+    }
+
+    updateEntry(id, data) {
+        this.entries = this.entries.map((entry) => {
+            if (entry.id === id) {
+                return data;
+            }
+            return entry;
+        });
+    }
+
+    isAuthorized(authnHeader) {
+        return authnHeader === this.correctAuthnHeader;
+    }
+
+    unauthorizedResponse() {
+        return this._response({ status: 401 });
+    }
+
+    match(method, pathname, params, postData) {
+        if (method === "GET" && pathname.endsWith("/clients")) {
+            return this._response({ jsonBody: [mockData.client] });
+        } else if (
+            method === "GET" &&
+            pathname.endsWith("/employees") &&
+            params.get("clientId") === mockData.client.Id.toString() &&
+            params.get("employeeLoginName") === correctUser
+        ) {
+            return this._response({ jsonBody: [mockData.employee] });
+        } else if (
+            method === "GET" &&
+            pathname.endsWith("/calculationPositions") &&
+            params.get("clientId") === mockData.client.Id.toString() &&
+            params.get("favoritesOnly") === "true"
+        ) {
+            return this._response({ jsonBody: [mockData.calculationPosition] });
+        } else if (
+            method === "GET" &&
+            pathname.endsWith("/billings/hours") &&
+            params.get("clientId") === mockData.client.Id.toString() &&
+            params.get("employeeId") === mockData.employee.Id.toString() &&
+            params.get("calculationPositionId") ===
+            mockData.calculationPosition.Id.toString()
+        ) {
+            return this._response({ jsonBody: this.entries });
+        } else if (method === "POST" && pathname.endsWith("/billings/hours")) {
+            this.addEntry({
+                id: this.entries.length,
+                Date: postData.Date,
+                Quantity: postData.Quantity,
+                Remark: postData.Remark,
+            });
+            return this._response({});
+        } else if (method === "PUT" && pathname.indexOf("/billings/hours") > -1) {
+            const splittedPath = pathname.split("/");
+            const id = parseInt(splittedPath[splittedPath.length - 1], 10);
+            this.updateEntry(id, {
+                id,
+                Date: postData.Date,
+                Quantity: postData.Quantity,
+                Remark: postData.Remark,
+            });
+            return this._response({});
+        } else {
+            return null;
+        }
+    }
+
+    _response({ status = 200, jsonBody = {} }) {
+        return {
+            status: status,
+            headers: { "Access-Control-Allow-Origin": "*" },
+            body: JSON.stringify(jsonBody),
+        };
+    }
+}
