+++ tests/test.js
@@ -1,181 +1,10 @@
 import { expect, test } from "@playwright/test";
-import md5 from "crypto-js/md5.js";
-
-class LoginPage {
-  constructor(page) {
-    this.page = page;
-  }
-
-  async logIn(username, password) {
-    await this.page.goto("/");
-    await this.page.locator("#username").fill(username);
-    await this.page.locator("#password").fill(password);
-    await this.page.locator("text=Sign in").click();
-  }
-}
-
-class TimeEntriesPage {
-  constructor(page) {
-    this.page = page;
-  }
-
-  async setFromTo(from, to) {
-    await this.page.locator('label:has-text("Show from:")').fill(from);
-    await this.page.locator('label:has-text("to:")').fill(to);
-  }
-
-  async addEntry(year, month, date, hours, description, useEnter = false) {
-    await this.page.locator('[placeholder="2022-01-01"]').click();
-    await this.page
-      .locator('[placeholder="2022-01-01"] + .picker >> select')
-      .first()
-      .selectOption({ label: month });
-    await this.page
-      .locator('[placeholder="2022-01-01"] + .picker >> select')
-      .nth(2)
-      .selectOption(year);
-    await this.page
-      .locator(`[placeholder="2022-01-01"] + .picker >> span >> text="${date}"`)
-      .nth(0)
-      .click();
-    await this.page.locator('[placeholder="2:15"]').fill(hours);
-    await this.page
-      .locator('[placeholder="Working the work…"]')
-      .fill(description);
-    if (useEnter) {
-      await this.page.keyboard.press("Enter");
-    } else {
-      await this.page.locator("button >> text=ADD").click();
-    }
-  }
-
-  async editEntry(year, month, date, hours, description, useEnter = false) {
-    await this.page.locator("text=Edit").nth(0).click();
-    await this.page.locator('[placeholder="2022-01-01"]').nth(1).click();
-    await this.page.locator("select").first().selectOption({ label: month });
-    await this.page.locator("select").nth(2).selectOption(year);
-    await this.page.locator(`span >> text="${date}"`).click();
-    await this.page.locator('[placeholder="2:15"]').nth(1).fill(hours);
-    await this.page
-      .locator('[placeholder="Working the work…"]')
-      .nth(1)
-      .fill(description);
-    if (useEnter) {
-      await this.page.keyboard.press("Enter");
-    } else {
-      await this.page.locator("button >> text=Save").click();
-    }
-  }
-}
+import LoginPage from "./TestHelper/LoginPage";
+import TimeEntriesPage from "./TestHelper/TimeEntriesPage";
+import TroiApiStub from "./TestHelper/TroiApiStub";
 
 const correctUser = "user.name";
 const correctPassword = "s3cr3t";
-
-const mockData = {
-  client: {
-    Name: "DigitalService GmbH des Bundes",
-    Id: 123,
-  },
-  employee: {
-    Id: 456,
-  },
-  calculationPosition: {
-    DisplayPath: "My Project",
-    Id: 789,
-  },
-};
-
-class TroiApiStub {
-  constructor() {
-    this.entries = [];
-
-    this.correctAuthnHeader = `Basic ${btoa(
-      `${correctUser}:${md5(correctPassword)}`
-    )}`;
-  }
-
-  addEntry(entry) {
-    this.entries.push(entry);
-  }
-
-  deleteEntry(id) {
-    this.entries = this.entries.filter((entry) => entry.id !== id);
-  }
-
-  updateEntry(id, data) {
-    this.entries = this.entries.map((entry) => {
-      if (entry.id === id) {
-        return data;
-      }
-      return entry;
-    });
-  }
-
-  isAuthorized(authnHeader) {
-    return authnHeader === this.correctAuthnHeader;
-  }
-
-  unauthorizedResponse() {
-    return this._response({ status: 401 });
-  }
-
-  match(method, pathname, params, postData) {
-    if (method === "GET" && pathname.endsWith("/clients")) {
-      return this._response({ jsonBody: [mockData.client] });
-    } else if (
-      method === "GET" &&
-      pathname.endsWith("/employees") &&
-      params.get("clientId") === mockData.client.Id.toString() &&
-      params.get("employeeLoginName") === correctUser
-    ) {
-      return this._response({ jsonBody: [mockData.employee] });
-    } else if (
-      method === "GET" &&
-      pathname.endsWith("/calculationPositions") &&
-      params.get("clientId") === mockData.client.Id.toString() &&
-      params.get("favoritesOnly") === "true"
-    ) {
-      return this._response({ jsonBody: [mockData.calculationPosition] });
-    } else if (
-      method === "GET" &&
-      pathname.endsWith("/billings/hours") &&
-      params.get("clientId") === mockData.client.Id.toString() &&
-      params.get("employeeId") === mockData.employee.Id.toString() &&
-      params.get("calculationPositionId") ===
-        mockData.calculationPosition.Id.toString()
-    ) {
-      return this._response({ jsonBody: this.entries });
-    } else if (method === "POST" && pathname.endsWith("/billings/hours")) {
-      this.addEntry({
-        id: this.entries.length,
-        Date: postData.Date,
-        Quantity: postData.Quantity,
-        Remark: postData.Remark,
-      });
-      return this._response({});
-    } else if (method === "PUT" && pathname.indexOf("/billings/hours") > -1) {
-      const splittedPath = pathname.split("/");
-      const id = parseInt(splittedPath[splittedPath.length - 1], 10);
-      this.updateEntry(id, {
-        id,
-        Date: postData.Date,
-        Quantity: postData.Quantity,
-        Remark: postData.Remark,
-      });
-      return this._response({});
-    } else {
-      return null;
-    }
-  }
-
-  _response({ status = 200, jsonBody = {} }) {
-    return {
-      status: status,
-      headers: { "Access-Control-Allow-Origin": "*" },
-      body: JSON.stringify(jsonBody),
-    };
-  }
-}
 
 let apiStub;
